Hackathon/AcademGuard/requirements.txt
Flask
Flask-MySQLdb
python-dotenv
